pyproject.toml
[project]
name = "youtube"
version = "0.1.0"
description = "YouTube 影片转静态阅读版投影片 Web 应用"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.20",
    "aiofiles>=24.1.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "yt-dlp==2026.2.21",
    "ffmpeg-python>=0.2.0",
    "Pillow>=11.0.0",
    "webvtt-py>=0.5.1",
    "deep-translator>=1.11.4",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.3",
    "sse-starlette>=2.2.1",
    "openai>=2.21.0",
    "quickjs>=1.19.4",
    "yt-dlp-ejs>=0.5.0",
    "openai-whisper>=20250625",
    "opencc-python-reimplemented>=0.1.7",
]
